add tests and fix price value types ti BigDecimal instead of double

+++ src/main/resources/db/migration/V5__ddl.create_product_table.sql
@@ -5,7 +5,7 @@ create table product
             primary key,
     description varchar(1000),
     short_description varchar(255),
-    price       double precision,
+    price       decimal(12,2),
     rating      double precision,
     title       varchar(255),
     category_id bigint
